backend/requirements.txt
backports-datetime-fromisoformat==2.0.3
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
